requirements.txt
prawcore==2.3.0
psaw==0.1.0
requests==2.26.0
update-checker==0.18.0
urllib3==1.26.6
websocket-client==1.2.1